feat(API Cleanup): rearranged connection, metadata and query backend
