+++ query-builder-backend/src/connection-manager/connection-manager.module.ts
@@ -1,9 +1,11 @@
 import { Module } from '@nestjs/common';
 import { ConnectionManagerService } from './connection-manager.service';
 import { JsonConverterService } from 'src/jsonConverter/jsonConverter.service';
+import { ConnectionManagerController } from './connection-manager.controller';
 
 @Module({
   providers: [ConnectionManagerService, JsonConverterService],
-  exports: [ConnectionManagerService]
+  exports: [ConnectionManagerService],
+  controllers: [ConnectionManagerController]
 })
 export class ConnectionManagerModule {}
