+++ query-builder-backend/src/db-metadata-handler/db-metadata-handler.controller.ts
@@ -1,4 +1,41 @@
-import { Controller } from '@nestjs/common';
+import { Body, Controller, Put } from '@nestjs/common';
+import { DbMetadataHandlerService } from './db-metadata-handler.service';
 
-@Controller('db-metadata-handler')
-export class DbMetadataHandlerController {}
+interface DatabaseCredentials {
+    host: string;
+    user: string;
+    password: string;
+}
+
+interface TableQuery {
+    credentials: DatabaseCredentials;
+    schema: string;
+}
+
+interface FieldQuery {
+    credentials: DatabaseCredentials;
+    schema: string;
+    table: string;
+}
+
+@Controller('metadata')
+export class DbMetadataHandlerController {
+
+    constructor(private readonly dbMetadataHandlerService: DbMetadataHandlerService) {}
+
+    @Put("schemas")
+    async getSchemaMetadata(@Body() credentials: DatabaseCredentials) {
+        return this.dbMetadataHandlerService.getSchemaMetadata(credentials);
+    }
+
+    @Put("tables")
+    async getTableMetadata(@Body() tableQuery: TableQuery) {
+        return this.dbMetadataHandlerService.getTableMetadata(tableQuery);
+    }
+
+    @Put("fields")
+    async getFieldMetadata(@Body() fieldQuery: FieldQuery){
+        return this.dbMetadataHandlerService.getFieldMetadata(fieldQuery);
+    }
+
+}
