+++ query-builder-backend/src/query-handler/query-handler.module.ts
@@ -1,9 +1,10 @@
 import { Module } from '@nestjs/common';
 import { QueryHandlerController } from './query-handler.controller';
 import { QueryHandlerService } from './query-handler.service';
+import { JsonConverterService } from 'src/jsonConverter/jsonConverter.service';
 
 @Module({
   controllers: [QueryHandlerController],
-  providers: [QueryHandlerService]
+  providers: [QueryHandlerService, JsonConverterService]
 })
 export class QueryHandlerModule {}
